# server/v2/dataservice/dataservice_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DataService V2 API",
    "description": "Data management service with FERPA compliance for scholarship platform",
    "version": "2.0.0-sprint2"
  },
  "paths": {
...
    "/api/v1/ledgers": {
      "post": {
        "tags": [
          "ledgers"
        ],
        "summary": "Create Ledger",
        "operationId": "create_ledger_api_v1_ledgers_post",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ledgers"
        ],
        "summary": "List Ledgers",
        "operationId": "list_ledgers_api_v1_ledgers_get",
        "parameters": [
          {
            "name": "reference_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference Type"
            }
          },
          {
            "name": "reference_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference Id"
            }
          },
          {
            "name": "is_balanced",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Balanced"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerResponse"
                  },
                  "title": "Response List Ledgers Api V1 Ledgers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledgers/reconcile": {
      "get": {
        "tags": [
          "ledgers"
        ],
        "summary": "Reconcile Ledger",
        "operationId": "reconcile_ledger_api_v1_ledgers_reconcile_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Trace ID to reconcile",
              "title": "Trace Id"
            },
            "description": "Trace ID to reconcile"
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledgers/{ledger_id}": {
      "get": {
        "tags": [
          "ledgers"
        ],
        "summary": "Get Ledger",
        "operationId": "get_ledger_api_v1_ledgers__ledger_id__get",
        "parameters": [
          {
            "name": "ledger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ledger Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledgers/{ledger_id}/validate": {
      "post": {
        "tags": [
          "ledgers"
        ],
        "summary": "Validate Ledger",
        "operationId": "validate_ledger_api_v1_ledgers__ledger_id__validate_post",
        "parameters": [
          {
            "name": "ledger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ledger Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountType": {
        "type": "string",
        "enum": [
          "asset",
          "liability",
          "equity",
          "revenue",
          "expense"
        ],
        "title": "AccountType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LedgerCreate": {
        "properties": {
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reference_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Type"
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryCreate"
            },
            "type": "array",
            "minItems": 2,
            "title": "Entries"
          },
          "extra_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Data"
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "LedgerCreate"
      },
      "LedgerEntryCreate": {
        "properties": {
          "account_type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "account_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Account Code"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Name"
          },
          "amount": {
            "type": "number",
            "title": "Amount",
            "description": "Debit positive, credit negative"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "USD"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "extra_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Data"
          }
        },
        "type": "object",
        "required": [
          "account_type",
          "account_code",
          "amount"
        ],
        "title": "LedgerEntryCreate"
      },
      "LedgerEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ledger_id": {
            "type": "string",
            "title": "Ledger Id"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type"
          },
          "account_code": {
            "type": "string",
            "title": "Account Code"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Name"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ledger_id",
          "account_type",
          "account_code",
          "amount",
          "currency",
          "created_at"
        ],
        "title": "LedgerEntryResponse"
      },
      "LedgerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reference_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Type"
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id"
          },
          "validated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated At"
          },
          "is_balanced": {
            "type": "boolean",
            "title": "Is Balanced"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryResponse"
            },
            "type": "array",
            "title": "Entries"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trace_id",
          "is_balanced",
          "entries",
          "created_at"
        ],
        "title": "LedgerResponse"
      },
...
      "ReconcileResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "ledger_count": {
            "type": "integer",
            "title": "Ledger Count"
          },
          "total_debits": {
            "type": "number",
            "title": "Total Debits"
          },
          "total_credits": {
            "type": "number",
            "title": "Total Credits"
          },
          "net_balance": {
            "type": "number",
            "title": "Net Balance"
          },
          "is_balanced": {
            "type": "boolean",
            "title": "Is Balanced"
          },
          "ledgers": {
            "items": {
              "$ref": "#/components/schemas/LedgerResponse"
            },
            "type": "array",
            "title": "Ledgers"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "ledger_count",
          "total_debits",
          "total_credits",
          "net_balance",
          "is_balanced",
          "ledgers"
        ],
        "title": "ReconcileResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}